Fixed favorite managment.
Before when the product favorite was set for same product ID but those two product objects where different the favorite was being set only on the object on which it was called the other object favorite wasn't being changed cause that was a different object. The solution was making favorites centralized. Now whenever a favorite is selected or loaded the same Favorite object will be returned, and thus if if that Favorite object is changed from one place in the other place it will be updated automatically and through INotifyPropertyChanged the user interface will also be updated.